// src/models/banner.d.ts
export interface BannerNewsModel {
  id: number;
  interest_id?: any;
  title: string;
  type: number;
  description: string;
  status: string;
  created_by: string;
  link?: any;
  image: string;
  image_mobile: string;
  created_at: number;
  updated_at: number;
  updated_by?: any;
  priority: string;
  color?: any;
  popup_image?: string;
}